refactor: extract game logic into hook
